feat: streaming loading dashboard page
