+++ app/ui/dashboard/revenue-chart.tsx
@@ -1,19 +1,16 @@
 import { generateYAxis } from '@/app/lib/utils'
 import { CalendarIcon } from '@heroicons/react/24/outline'
 import { lusitana } from '@/app/ui/fonts'
-import { Revenue } from '@/app/lib/definitions'
+import { fetchRevenue } from '@/app/lib/data'
 
-export default async function RevenueChart({
-  revenue,
-}: {
-  revenue: Revenue[]
-}) {
+export default async function RevenueChart() {
+  const revenue = await fetchRevenue()
+
   const chartHeight = 350
-
-  const { yAxisLabels, topLabel } = generateYAxis(revenue);
+  const { yAxisLabels, topLabel } = generateYAxis(revenue)
 
   if (!revenue || revenue.length === 0) {
-    return <p className="mt-4 text-gray-400">No data available.</p>;
+    return <p className="mt-4 text-gray-400">No data available.</p>
   }
 
   return (
@@ -21,7 +18,6 @@ export default async function RevenueChart({
       <h2 className={`${lusitana.className} mb-4 text-xl md:text-2xl`}>
         Recent Revenue
       </h2>
-      {/* NOTE: comment in this code when you get to this point in the course */}
 
       <div className="rounded-xl bg-gray-50 p-4">
         <div className="sm:grid-cols-13 mt-0 grid grid-cols-12 items-end gap-2 rounded-md bg-white p-4 md:gap-4">
